[Fuzzing] updates for oss-fuzz Ideal integration
