+++ tests/fuzzing_seed_corpus/fuzz_sonmp_seed_corpus/sonmp.0.raw
